Rename webapp_photo_luminescence to dawa_trpl
