+++ src/dawa_trpl/__main__.py
@@ -1,4 +1,4 @@
-from webapp_photo_luminescence import app
+from dawa_trpl import app
 
 if __name__ == "__main__":
     app.run_server(debug=True)
