+++ src/dawa_trpl/components/tabs/h_figure_tab/process.py
