+++ src/dawa_trpl/components/tabs/v_figure_tab/__init__.py
@@ -9,10 +9,10 @@ import pandas as pd
 import plotly.graph_objects as go
 from dash import dcc
 
-from webapp_photo_luminescence import data_system as ds
-from webapp_photo_luminescence.components import upload_bar
-from webapp_photo_luminescence.components.tabs import common
-from webapp_photo_luminescence.components.tabs.v_figure_tab import process
+from dawa_trpl import data_system as ds
+from dawa_trpl.components import upload_bar
+from dawa_trpl.components.tabs import common
+from dawa_trpl.components.tabs.v_figure_tab import process
 
 wavelength_slider = dcc.RangeSlider(
     id="v-wavelength-slider",
