+++ src/dawa_trpl/components/tabs/v_figure_tab/process.py
