+++ tests/unit/test_components_tabs_h_figure_tab_process.py
@@ -4,8 +4,8 @@ import plotly.graph_objects as go
 import pytest
 import tlab_analysis.photo_luminescence as pl
 
+from dawa_trpl.components.tabs.h_figure_tab import process
 from tests import IMGDIR
-from webapp_photo_luminescence.components.tabs.h_figure_tab import process
 
 
 @pytest.fixture(scope="module")
