+++ tests/unit/test_components_tabs_v_figure_tab.py
@@ -9,8 +9,8 @@ import plotly.graph_objects as go
 import pytest
 import tlab_analysis.photo_luminescence as pl
 
+from dawa_trpl.components.tabs import v_figure_tab
 from tests import IMGDIR
-from webapp_photo_luminescence.components.tabs import v_figure_tab
 
 
 @pytest.fixture()
@@ -32,9 +32,7 @@ def wrs() -> list[pl.TimeResolved]:
 
 @pytest.fixture()
 def process_mock() -> abc.Generator[mock.Mock, None, None]:
-    with mock.patch(
-        "webapp_photo_luminescence.components.tabs.v_figure_tab.process"
-    ) as m:
+    with mock.patch("dawa_trpl.components.tabs.v_figure_tab.process") as m:
         yield m
 
 
@@ -44,7 +42,7 @@ def ds_mock(
     selected_items: list[str],
     upload_dir: str,
 ) -> abc.Generator[mock.Mock, None, None]:
-    with mock.patch("webapp_photo_luminescence.components.tabs.v_figure_tab.ds") as m:
+    with mock.patch("dawa_trpl.components.tabs.v_figure_tab.ds") as m:
         m.load_wavelength_resolveds.return_value = wrs
         m.get_existing_item_filepaths.return_value = [
             os.path.join(upload_dir, item) for item in selected_items
