+++ tests/unit/test_components_tabs_v_figure_tab_process.py
@@ -4,9 +4,9 @@ import plotly.graph_objects as go
 import pytest
 import tlab_analysis.photo_luminescence as pl
 
+from dawa_trpl import data_system as ds
+from dawa_trpl.components.tabs.v_figure_tab import process
 from tests import IMGDIR
-from webapp_photo_luminescence import data_system as ds
-from webapp_photo_luminescence.components.tabs.v_figure_tab import process
 
 
 @pytest.fixture(scope="module")
